consolidate audio utils
