+++ src/pages/index.html
@@ -396,7 +396,7 @@
     import { findMusicalLoop, findDownbeatPhase } from '/src/core/librosa-downbeat.js';
     
     // Audio utilities
-    import { computeRMS, computePeak, computeZeroCrossingRate } from '/src/core/audio-utils.js';
+    import { computeRMS, computePeak, computeZeroCrossingRate } from '/src/utils/audio-utils.js';
     
     // Dynamic zero crossing for clean loops
     import { DynamicZeroCrossing } from '/src/core/dynamic-zero-crossing.js';
